+++ src/wai/common/meta/__init__.py
@@ -1,0 +1,4 @@
+"""
+Package for meta-programming constructs.
+"""
+from ._metadata import with_metadata, get_metadata, has_metadata