+++ src/wai/common/meta/_metadata.py
@@ -1,0 +1,48 @@
+from typing import Any
+
+from .._typing import T
+
+META_DATA_KEY: str = "__metadata"
+
+
+def with_metadata(obj: T, key: str, value: Any) -> T:
+    """
+    Adds meta-data to an object.
+
+    :param obj:     The object to add meta-data to.
+    :param key:     The key to store the meta-data under.
+    :param value:   The meta-data value to store.
+    :return:        obj.
+    """
+    # Create the meta-data map
+    if META_DATA_KEY not in obj.__dict__:
+        obj.__dict__[META_DATA_KEY] = {}
+
+    # Put this mapping in the map
+    obj.__dict__[META_DATA_KEY][key] = value
+
+    return obj
+
+
+def get_metadata(obj: Any, key: str) -> Any:
+    """
+    Gets the meta-data from an object.
+
+    :param obj:     The object to get meta-data from.
+    :param key:     The meta-data key to extract from.
+    :return:        The value of the meta-data.
+    """
+    return obj.__dict__[META_DATA_KEY][key]
+
+
+def has_metadata(obj: Any, key: str) -> bool:
+    """
+    Checks if the given object has meta-data associated with
+    the given key.
+
+    :param obj:     The object to check.
+    :param key:     The meta-data key.
+    :return:        True if there is meta-data associated with the
+                    given key, False if not.
+    """
+    return META_DATA_KEY in obj.__dict__ and key in obj.__dict__[META_DATA_KEY]