+++ test/__init__.py
@@ -1,1 +1,2 @@
 from ._IntervalTest import IntervalTest
+from ._TwoWayDictTest import TwoWayDictTest